Allow the Link Checker API tests to pass
The Database Cleaner gem was failing because it won't work on remote databases by default.

Database Cleaner considers a remote database as any database which doesn't have a hostname of `localhost` or `127.0.0.1`. In this context, the Postgres database counts as remote because its hostname is `postgres-9.6`.

This change adds an environment variable so that Database Cleaner will work within this docker environment.