Update Content Publisher to use PostgreSQL 13
This is the version we are planning to upgrade PostgreSQL to in production, so we need to update our development environment too.

[Trello card][1]

Note: the test suite for this app doesn't currently pass in the docker environment. But the test failures seem to be unrelated to the database upgrade, as evidenced by the fact that [they pass in CI][2] with the same database upgrade.

We have a separate [Trello card][3] to resolve the failing tests in the docker environment.

[1]: https://trello.com/c/zVdVJqFM/
[2]: https://github.com/alphagov/content-publisher/pull/2431
[3]: https://trello.com/c/QZOaYQCH/