Update `replicate-mongodb.sh` script for Router database
This commit updates the `replicate-mongodb.sh` script to work for the `router` and `draft-router` databases. Specifically, Router uses MongoDB 2.6, but the replication script previously made no provision for this and so wouldn't have worked for these databases.

When restoring a back-up of either `router` or `draft-router` using MongoDB 2.6, a `sleep 60` has been added to allow for the replica set to properly initialise, as described by MongoDB when you run `rs.initiate()`; whilst I can't find any documentation related to this on the MongoDB docs, there are references to it elsewhere (see [here](https://medium.com/@ManagedKube/deploy-a-mongodb-cluster-in-steps-9-using-docker-49205e231319) and [here](https://www.ceondo.com/notes/mongodb/)).