Define replica set config for MongoDB 2.6
This commit adds a definition for the replica set config to use for MongoDB 2.6. Previously in this repo we initialised a replica set without specifying any config, which has the effect of creating a config using the container id of said container running MongoDB 2.6. The issue here is that this replica set config is persisted to the `govuk-docker_mongo-2.6` volume, meaning that when the current container is eventually removed and a new container created to run MongoDB 2.6 the replica set won't be properly provisioned as it references the wrong container id.

To get around this problem in govuk-docker, this commit specifies the replica set config to use during initialisation and uses the name of the container to route to, rather than the container id; through this change, the replica set config can remain persisted in the associated volume whilst containers come and go, all whilst enabling these new instances to take part in the replica set. This isn't a problem with MongoDB 2.6 in production as we're not running it in containers.

For clarity, the replica set is used by Router to poll for updates to Routes, using the `optime` property from `rs.status()` to determine if it holds an up-to-date copy of routes; for more context [see this commit](https://github.com/alphagov/router/commit/678ef634657885e4eb552954ef89dcc29b47193c) and [RFC](https://github.com/alphagov/govuk-rfcs/blob/main/rfc-135-updating-routes-in-router.md).