Remove Postgres 9.6 and MySQL 5.5 Databases
This removes the configuration for Postgres 9.6 and MySQL 5.5 as no
applications use these anymore.

Developers can perform some clean-up tasks to recover the space used by
these, if they wish. This is probably only of particular value if you
have cloned large databases.

The most space saving measure would be removing the volumes:

$ docker volume rm govuk-docker_postgres-9.6 govuk-docker_mysql-5.5

The containers and images can be removed to with:

Stop and remove the containers (these may error if all the containers
are already stopped):

$ docker container rm $(docker container ls -aq --filter "ancestor=mysql:5.5")
$ docker container rm $(docker container ls -aq --filter "ancestor=postgres:9.6")

Remove the images:

$ docker image rm postgres:9.6 mysql:5.5