Fix `bin/doctor` output when dnsmasq isn't running
Without this fix, the `bin/doctor` script output looks broken:

```
Checking dnsmasq
usage: grep [-abcdDEFGHhIiJLlMmnOopqRSsUVvwXxZz] [-A num] [-B num] [-C[num]]
	[-e pattern] [-f file] [--binary-files=value] [--color=when]
	[--context[=num]] [--directories=action] [--label] [--line-buffered]
	[--null] [pattern] [file ...]
```

This is because `pgrep dnsmasq` was returning an empty result when
dnsmasq wasn't running properly, and we're attempting to `grep`
that empty result, so we're effectively calling `grep` with no
arguments, and thus we're seeing `grep`'s help message here.

The fix in this commit detects when `pgrep` gives an empty output,
and in that case will pass `dnsmasq_not_running` to `grep`, so
that `grep` doesn't think we're calling it with no arguments. I
did first try setting a default argument of `""`, but that doesn't
fix the issue. I needed to pick a default argument that wasn't
likely to exist on the list of processes, so
`'#{service_name}_not_running'` seemed a suitable value for that.