Make `dig` time out after 1 second
The default timeout is 5 seconds, and the default number of
retries is 3, meaning that when `dnsmasq` is not resolving, we're
left waiting for 15 seconds to find that out.

As this is a local DNS change, the `dig` should be more or less
instantaneous - and should not need any retries - so it seems
sensible to set the timeout to 1 second and set it to only try to
connect once. This greatly speeds up the `bin/doctor` script for
developers who don't yet have `dnsmasq` working.