Add project for Smokey
This avoid adding a bunch of guidance to the README about how to
install Rbenv and Chromedriver, which we now expect to be managed
by the system package manager [^1]. Note that:

- "NO_SANDBOX" is similar to the setting we have in the base image
[^2], but specific to Smokey's test setup [^3].

- "shm_size" is necessary to avoid Chrome crashing on certain tests
[^4]. "watch"ing /dev/shm locally, the highest I saw was 81Mb.

Running the tests through Docker is slightly slower - locally the
tests take around 20s longer, with a baseline of 3 mins. This is
similar to the tradeoff we have for other apps.

[^1]: https://github.com/alphagov/govuk_test/commit/50ed673b1bd278a51f45eb241d7949918ce77d8e
[^2]: https://github.com/alphagov/govuk-docker/blob/32449a69301f20770e9606f22467658efa948eb0/Dockerfile.govuk-base#L12
[^3]: https://github.com/alphagov/smokey/blob/2017266b906ec7f8f96fb164cf3d8ba03c587296/features/support/env.rb#L40
[^4]: https://stackoverflow.com/questions/53902507/unknown-error-session-deleted-because-of-page-crash-from-unknown-error-cannot