Run linting as part of overall test recipe
This matches the current behaviour of CI [^1].

Making this change will help simplify the way we run tests locally.
I've split out the non-Ruby test commands into a new "test-scripts"
recipe as part of this - so that we can reuse it elsewhere.

Normally we would do linting as part of the default "make" task,
but that's not appropriate here as we also run make commands to set
up individual projects; having the default "make" run tests would be
confusing for someone who's just trying to use GOV.UK Docker.

I think it's reasonable to include linting as part of the "test" recipe,
as we generally want to run both if we're changing Ruby code.

[^1]: https://github.com/alphagov/govuk-docker/blob/main/.github/workflows/ci.yml