Add govuk-docker project to run repo Ruby commands
This is a bit meta but much easier than forcing someone to create
their own Ruby environment for this purpose.

This leaves bin/doctor as the only part of the repo that requires
a local Ruby environment, so I've moved the installation guidance
to the relevant part of the troubleshoot manual.

Since some of the testing commands assume local file paths or run
Docker itself, we can't run all the "tests" inside a container, so
I've added separate recipes / commands to run the tests on CI and
locally. I'm not worried about them diverging.