Install `less` for scrolling in Rails consoles
This installs `less` which is a common utility on Linux systems used for scrolling up/down within large documents. But unfortunately it doesn't exist in the base image we're using.

In Rails applications, `rails console` and `binding.pry` commands both open a Pry session. When debugging something which is too large to fit in the terminal window, Pry will attempt to paginate the output using `less`. If that's not available on the local system, it'll fall back to a 'SimplePager' implementation which is weird and not nice to use.

To make the development environment more predictable, I've installed `less` so that we get pagination which is nice to use.