Remove ASSET_HOST env vars
These env vars are unnecessary as they just set the asset host to be the
same as the virtual host and don't impact the app.

They are being removed as I'm going to remove the ASSET_HOST env var as
updates to the Content Security Policy are going to require the assets
to be served from self - as is the case in production.

It's possible these asset hosts were added to enable proxying to the
apps through router. However, when I tried to replicate this
configuration there were numerous issues that led me to conclude that
no-one was used this in a long time - principally it is that router
can't load routes without a mongo with a replica set.

Should we wish to re-enable to ability to proxy to apps via router then
we should do this via nginx configuration. This could be done by
mounting a vhost config file for vhosts in the nginx-proxy container
configuration e.g:

```
    volumes:
      - ./nginx-router-assets.conf:/etc/nginx/vhost.d/www.dev.gov.uk
      - ./nginx-router-assets.conf:/etc/nginx/vhost.d/router.dev.gov.uk
      - ./nginx-router-assets.conf:/etc/nginx/vhost.d/www-origin.dev.gov.uk
```

With a nginx-router-assets.conf being something like:

```
location /assets/frontend/ {
  proxy_pass http://frontend.dev.gov.uk;
}

location /assets/static/ {
  proxy_pass http://static.dev.gov.uk;
}

...
```