Proxy static in frontend apps
This changes the configuration of the *-app stacks of frontend apps so
that they proxy any requests to /assets/static to static (as per the
*-app-live configuration).

It removes the ASSET_HOST configuration from static so that static won't
set a host.

The motivation for this change is an update to the Content Security
Policy [1] which removes the usage of hosts for assets.

[1]: https://github.com/alphagov/govuk_app_config/pull/274