Update replicate-mongodb.sh to use govuk-env-sync backups
This script was using backups that are generated specifically for
local Docker development under /mongodb/daily/ in the S3 bucket,
whereas all other backups use the govuk-env-sync process.

The following have already been migrated:
- maslow_production (DocumentDB)
- short_url_manager_production (DocumentDB)
- govuk_assets_production (DocumentDB)
- travel_advice_publisher_production (DocumentDB)
- govuk_content_production (DocumentDB)
- imminence_production (PostgreSQL)

These apps are the last ones remaining on MongoDB and will be migrated
to DocumentDB or PostgreSQL. authenticating-proxy has not been included
as is in the process of being migrated - see https://trello.com/c/lSpntlfk/81-mongo-26-has-reached-end-of-life
for more information.