Remove references to govuk-content-schemas
Content schemas have now been moved into publishing api, so this removes the docker compose, make file, references to it in documentation, and any reference to it as a dependency for another application.

The only two applications with explicit dependencies on govuk-content-schemas in their makefile are publishing api and content store. Publishing api now contains the schemas, so doesn't need it as a dependency. Content store's makefile has been modified to add publishing-api as a dependency.