Update test to reflect moving of schemas into publishing api
Govuk-content-schemas have been retired and moved into publishing api. The only application that depends on publishing-api in the makefiles at the moment is content-store, which depends on publishing-api solely for its schemas. Therefore, it's OK that there's a difference here between the dependent projects and the applications it needs running.