Make branch_checks.sh respect GOVUK_ROOT_DIR
The branch_checks.sh script currently has the root directory for GOV.UK
repos hardcoded as `$HOME/govuk` and doesn't use the `GOVUK_ROOT_DIR`
environment variable that the rest of the project does. This means
`make [anyproject]` will fail on configurations where the user has a
custom `GOVUK_ROOT_DIR` set up.

This sets the root directory to the value of `$GOVUK_ROOT_DIR` if set,
or the existing default otherwise.