Add ability to run some frontends apps against integration data
This allows us to run Collections, Finder Frontend and Government
Frontend against the data present in the integration environment.

We sometimes use integration to test the running of rake tasks or data
migrations, so it is useful to have a local way of testing frontend
changes against that data.

The integration app can be started using:

```
govuk-docker-up app-integration
```