Pin Dockerfiles to buildpacks-dep:bullseye
I ran into a problem this morning when running `make publishing-api`. It
was failing with the following error:

```
LoadError: libssl.so.1.1: cannot open shared object file: No such file or directory -
/root/.rbenv/versions/3.1.2/lib/ruby/3.1.0/x86_64-linux/openssl.so
/govuk/publishing-api/config/application.rb:3:in `<main>'
/govuk/publishing-api/Rakefile:4:in `<main>'
(See full trace by running task with --trace)
make: ***
[/home/chrisroos/govuk/govuk-docker/projects/publishing-api/Makefile:2:publishing-api] Error 1
```

By not specifying a tag of the buildpack-deps image we were implicitly
using `latest` which, as far as I can tell[1], is the same as `stable`.
The `stable` tag was recently updated[2] from Debian 11 (Bullseye) to 12
(Bookwork). The upgrade from Debian 11 to 12 brings an upgrade of
OpenSSL from 1.1.1n to 3.0.9 and this appears to be the cause of the
problems I was seeing.

By specifying the `bullseye` tag I believe we should be retaining the
behaviour from before the change to the `stable` tag.

Prior to this change running `make publishing-api`, `make asset-manager`
and `make whitehall` all failed with similar SSL problems. After this
change they all build successfully.

[1]: https://github.com/docker-library/buildpack-deps/blob/de709c6cb259198ef49cef48f395eac92e2288f2/generate-stackbrew-library.sh#L79
[2]: https://github.com/docker-library/buildpack-deps/commit/9109c788a9e2d5c2103fc74e3ebbf713b459a3fb