Fix replicate-mongodb.sh to work on linux
The `\d` in the regexp passed to grep isn't supported in the GNU version
of grep. I understand that the equivalent `[[:digit:]]` should work in
all implementations.

The `-z` option to `tar` results in the following error on linux:

```
gzip: stdin: not in gzip format
tar: Child died with signal 13
tar: Error is not recoverable: exiting now
```

The `-z` option filters the output through gzip but I don't believe this
is necessary because we're explicitly `gunzip`ing the dump before
passing it to `tar`.

With these two changes the script runs without error on my linux
development environment.