Fix data replication script for search-api
After a lot of trial-and-error and investigation where we saw the elasticesearch config YAML
file reporting as a directory! And what appeared to be a user (docker image user) permissions
issue. Finally, I found that by moving the path.repo config option into the dummy config
file it resolved the issue.