Replicate data from staging, not integration
Since the env sync was ported over to govuk-helm-charts, not all
apps have backups saved to s3 on integration (e.g. note the lack
of `backup` operation [here](https://github.com/alphagov/govuk-helm-charts/blob/cae5e10baf0f67423bbebf66454da737da787d37/charts/db-backup/values.yaml#L767-L773)).
This means that the data replication scripts fail, as they're
looking for a bucket that does not exist.

In contrast, all data is backed up to s3 on staging, so it should
be safe to pull from there by default. The only downside is that
integration access is given to devs by default, but that access to
staging and production required Production Admin access. This is
something that can hopefully be ironed out in the move to using
govuk-helm-charts for local development, over the coming months.
For now it seems better to fix the script for prod-admin users
than to have it broken for everyone.

NB I've also removed the unnecessary duplication of replication
examples where one used a readonly role and one used a poweruser
role. We should use principle of least privilege, so only show
the readonly example.