Add search autocomplete API configuration
This adds `SEARCH_AUTOCOMPLETE_API_URL` environment variable with an
appropriate value to all apps that show an autocompletable search field.

Note that for `frontend` and `static`, we're using the production
endpoint even for the local `app` stack, as being able to test these
with a local `search-api-v2` isn't much value and adding a local
dependency between the apps adds setup friction.