Add publishing-api dependency to whitehall-lite
The lite app is the one created when running `make whitehall`.
In #837 we automatically run the 'taxonomy building' rake task
as part of the `make` command, but that rake task depends on
Publishing API being up and running, so the make was failing
because publishing-api hadn't been listed as a dependency.