On GitHub Actions dependency updates: convert tags to commit hashes
We want all of our GitHub Actions actions dependencies to be pinned to a
specific commit hash, but Dependabot will only pin to a tag. To work around
this we introduce this workflow, which runs whenever Dependabot raises a
relevant PR, and uses a tool called pinact to convert the tags to the commit id
to which they point at that moment in time.