Update GOV.UK Chat's docker compose to include multiple workers
We're implementing sidekiq in two separate processes to ensure we process
users questions in a high priority queue with more workers and resources.
And a low priority queue with less workers and resources which handles
post answer composition background jobs which aren't time-sensitive.

This adds in the necessary configuration for the new worker processes.