Add new search-api-v2-beta-features
At the moment this app is only be used to run evaluation tasks and therefore
doesn't need a web server.

There is no need for an app stack. The tests for the "app" stack check
that the service depends on nginx-proxy and that a default host is set-up,
neither of which are currently needed for this project.
A `task-runner` stack is being used to test the rake tasks locally instead.
This matches the configuration in [search-api-v2].

[search-api-v2](https://github.com/alphagov/govuk-docker/blob/7d6bb91ad952c5f797027832488a57ff0f87ad9a/projects/search-api-v2/docker-compose.yml#L69)

Add GOOGLE_CLOUD_PROJECT_ID_NUMBER

Change name to search-api-v2-beta-features