Add GOOGLE_CLOUD_PROJECT_ID_NUMBER env var
GCP has two ways of referencing a project. Either with a human readable
id (GOOGLE_CLOUD_PROJECT_ID) or with a number. Most API calls to the
Google REST client accept the human readable id. However we have found
one that does not.

This adds the project id number as a separate env var so that we can
decide which value to use for each api call.