Prevent orphaned carrierwave files
Over time, files would build up in the `carrierwave-tmp` directory when tests requiring carrierwave uploads were run.

These files were not required beyond the duration of the test.

Rubocop would become increasingly slow at scanning the project because it was not configured to ignore the carrierwave-tmp directory.

Adding a volume for the carrierwave-tmp directory enables the test process to delete the files once the tests have completed because they are not owned by the external user.