Remove living document references to Content Publisher
I've replaced references to Content Publisher with Collections
Publisher (partly to only replace one word, and partly because it
is more reliably interchangable than `whitehall`, which sometimes
has to be referred to as `whitehall-admin`).

I haven't updated references in the ADRs as they were snapshots
of the thinking at the time, and link to sha URLs which should
continue to work after this PR is merged.